backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
websockets>=12.0
chromadb>=0.4.0
pdfplumber>=0.10.0
sentence-transformers>=2.2.0
reportlab>=4.0
